requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.3.0
openfermion>=1.3.0
pyyaml>=5.4.0
